[ScheduleDevelopProejct] Comment CRUD 수정
1. Comment Create, Update, Delete 시 SessionId 확인 후 진행하도록 수정

+++ src/main/java/com/example/scheduledevelopproject/service/Impl/CommentServiceImpl.java
@@ -5,14 +5,20 @@ import com.example.scheduledevelopproject.dto.CommentResponseDto;
 import com.example.scheduledevelopproject.entity.Comment;
 import com.example.scheduledevelopproject.entity.Schedule;
 import com.example.scheduledevelopproject.entity.User;
+import com.example.scheduledevelopproject.exception.CustomException;
+import com.example.scheduledevelopproject.exception.ErrorCode;
 import com.example.scheduledevelopproject.repository.CommentRepository;
 import com.example.scheduledevelopproject.repository.ScheduleRepositroy;
 import com.example.scheduledevelopproject.repository.UserRepository;
 import com.example.scheduledevelopproject.service.CommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import static com.example.scheduledevelopproject.exception.ErrorCode.INVALID_USER_NAME;
 
 @Service
 @RequiredArgsConstructor
@@ -21,11 +27,16 @@ public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
     private final ScheduleRepositroy scheduleRepositroy;
+    private Long sessionId;
 
     @Override
-    public CommentResponseDto createUser(Long scheduleId, CommentRequestDto requestDto) {
+    public CommentResponseDto createUser(Long scheduleId, CommentRequestDto requestDto, Long sessionId) {
+        User user = userRepository.findByIdOrElseThrow(sessionId);
+        if(!requestDto.getUsername().equals(user.getUsername())){
+            throw new CustomException(INVALID_USER_NAME);
+        }
+
         Schedule schedule = scheduleRepositroy.findByIdOrElseThrow(scheduleId);
-        User user = userRepository.findUserByUsernameOrElseThrow(requestDto.getUsername());
         //Comment createComment = new Comment(requestDto.getComment());
         Comment createdComment = Comment.builder()
                 .content(requestDto.getComment())
@@ -39,34 +50,70 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public CommentResponseDto findByIdComment(Long scheduleId, Long commentsId) {
         Schedule schedule = scheduleRepositroy.findByIdOrElseThrow(scheduleId);
-        User user = userRepository.findUserByUsernameOrElseThrow(schedule.getUser().getUsername());
-        Comment findComment = commentRepository.findByIdOrElseThrow(commentsId);
-        return new CommentResponseDto(findComment);
-    }
-
-    @Override
-    public List<CommentResponseDto> findAllComment() {
-        return commentRepository.findAll()
+        Comment findComment = schedule.getComments()
                 .stream()
-                .map(CommentResponseDto::toDto)
-                .toList();
-    }
-
-    @Override
-    public CommentResponseDto updateComment(Long schedulesId, Long commentsId, CommentRequestDto requestDto) {
-        Comment findComment = commentRepository.findByIdOrElseThrow(commentsId);
-
-        Schedule schedule = scheduleRepositroy.findByIdOrElseThrow(findComment.getId());
-        User user = userRepository.findUserByUsernameOrElseThrow(requestDto.getUsername());
-        findComment.setContent(requestDto.getComment());
-        Comment savedComment = commentRepository.save(findComment);
+                .filter(comment -> comment.getId().equals(commentsId))
+                .findFirst()
+                .orElseThrow(
+                        ()-> new CustomException(ErrorCode.COMMENT_NOT_FOUND));
 
         return new CommentResponseDto(findComment);
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public Page<CommentResponseDto> findByScheduleId(Long scheduleId, int page, int size) {
+        if(page > 0){
+            page--;
+        }
+
+        Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "modifiedAt"));
+        Page<Comment> commentsPage = commentRepository.findByScheduleId(scheduleId, pageable);
+
+        // Comment -> CommentResponseDto 매핑
+        return commentsPage.map(CommentResponseDto::new);
+
+    }
+
+    @Override
+    public Page<CommentResponseDto> findAllComment(int page, int size) {
+        if(page > 0){
+             page--;
+        }
+
+        Pageable pageable = PageRequest.of(page,size, Sort.by(Sort.Order.asc("schedule_id")));
+        Page<Comment> comments = commentRepository.findAll(pageable);
+
+        return comments.map(CommentResponseDto::new);
+    }
+
+    @Override
+    public CommentResponseDto updateComment(Long schedulesId, Long commentsId, CommentRequestDto requestDto, Long sessionId) {
+        User user = userRepository.findByIdOrElseThrow(sessionId);
+        if(!requestDto.getUsername().equals(user.getUsername())){
+            throw new CustomException(INVALID_USER_NAME);
+        }
+
+        Schedule schedule = scheduleRepositroy.findByIdOrElseThrow(schedulesId);
+        //Comment createComment = new Comment(requestDto.getComment());
+        Comment createdComment = Comment.builder()
+                .content(requestDto.getComment())
+                .user(user)
+                .schedule(schedule)
+                .build();
+        Comment savedComment = commentRepository.save(createdComment);
+        return new CommentResponseDto(savedComment);
+    }
+
+    @Override
+    public void deleteComment(Long commentId,Long sessionId) {
+        User user = userRepository.findByIdOrElseThrow(sessionId);
         Comment comment = commentRepository.findByIdOrElseThrow(commentId);
+        if(!comment.getUser().getUsername().equals(user.getUsername())){
+            throw new CustomException(INVALID_USER_NAME);
+        }
+
         commentRepository.delete(comment);
     }
+
+
 }
